set input exp ang as float type

+++ Self_Balance_Vehicle/inc/AttitudeController.h
@@ -4,7 +4,7 @@
 #include "pidController.h"
 #include "AttitudeEstimator.h"
 
-void AttitudeControlLoop(uint8_t ExpAngle, uint8_t ControllerEnable);
+void AttitudeControlLoop(float ExpAngle, uint8_t ControllerEnable);
 void GetAttitudeControllerOutput(float *Left, float *Right);
 
 #endif /* __ATTITUDECONTROLLER_H */
